# src/api-docs/swagger.yaml
openapi: 3.1.0
...
info:
    title: Sentinel BankServer API
    version: 1.0.0
    summary: Encrypted-at-the-application-layer banking API.
    description: |
        Reference documentation for the BankServer HTTP surface. Every operation
        is grouped under one of the 18 numbered flow tags so the call order is
        self-evident. Each operation also carries:

          * `description` — markdown including a bulleted **Prerequisites** list
            of operationIds that must run before this one.
          * `x-call-order` — an integer that orders operations *within* the flow
            tag. Tools that respect extension fields can sort by this.
          * `x-depends-on` — machine-readable mirror of the prerequisites list.
          * `x-decrypted-request` / `x-decrypted-response` — for encrypted ops,
            the JSON shape carried inside the inner envelope's `data` field
            (i.e. the operation's actual payload). The wire-level
            `requestBody` / `responses` schemas keep the literal wire envelope
            (`{ sessionId, payload }` in / `{ payload }` out); see
            **Cryptographic envelope** below for the inner envelope.

        ## Cryptographic envelope

        After the bootstrap handshake (flow `01-Bootstrap`), every encrypted
        endpoint uses **two nested envelopes**:

        **1. Wire envelope** — what travels over HTTP:

        ```json
        // Request
...

        // Response
...
        ```

...

...

        ```json
...
        ```

...

...

...

...

...

...

...
servers:
...
security:
    - SessionId: []
...
components:
    parameters:
        SessionIdHeader:
            name: x-session-id
            in: header
            required: true
            description: UUID of the encrypted session (must match `sessionId` in the body).
            schema:
                type: string
                format: uuid
...
    securitySchemes:
        SessionId:
            type: apiKey
            in: header
            name: x-session-id
            description: |
                UUID identifying the encrypted session created by the bootstrap
                handshake. Required on every encrypted endpoint.
        ActionToken:
            type: apiKey
            in: header
            name: x-action-token
            description: |
                One-shot token minted by `postWebauthnAuthenticationVerify` and
                bound to a specific `(action, paramsHash, sessionId)` tuple.
                Required by every step-up gated operation.
        OtpToken:
            type: apiKey
            in: header
            name: x-otp-token
            description: |
                One-shot token minted by `postIdentityOtpVerify` and bound
                to a specific `(action, paramsHash, sessionId)` tuple.
                Required ALONGSIDE `ActionToken` by every OTP-gated route
                (passkey.add, passkey.revoke, password.change, session.wipe,
                admin.recovery).
    schemas:
...
        EnvelopedRequest:
            type: object
            required: [sessionId, payload]
            properties:
                sessionId:
                    type: string
                    format: uuid
                    description: Same UUID sent in the `x-session-id` header.
                payload:
                    type: string
                    description: |
                        Base64-encoded AES-256-GCM ciphertext (with
                        `sessionId` as AAD) of an inner
                        [`DecryptedEnvelope`](#/components/schemas/DecryptedEnvelope).
            description: |
                Generic wire shape of every encrypted POST body. After
                AES-GCM decryption, `payload` is a
                [`DecryptedEnvelope`](#/components/schemas/DecryptedEnvelope)
                — the operation's actual JSON sits in its `data` field, with
                shape declared per-operation by `x-decrypted-request`.
        EnvelopedResponse:
            type: object
            required: [payload]
            properties:
                payload:
                    type: string
                    description: |
                        Base64-encoded AES-256-GCM ciphertext (with
                        `sessionId` as AAD) of an inner
                        [`DecryptedEnvelope`](#/components/schemas/DecryptedEnvelope).
            description: |
                Generic wire shape of every encrypted success response. After
                AES-GCM decryption, `payload` is a
                [`DecryptedEnvelope`](#/components/schemas/DecryptedEnvelope)
                — the operation's actual JSON sits in its `data` field, with
                shape declared per-operation by `x-decrypted-response`.
...
        Role:
            type: string
            enum: [customer, admin]
...
        User:
            type: object
            required: [id, username, role]
            properties:
                id:
                    type: string
                    format: uuid
                username:
                    type: string
                role:
                    $ref: "#/components/schemas/Role"
...
        PasskeyOptionsInput:
            type: object
            required: [sessionId, nonce]
            properties:
                sessionId:
                    type: string
                    format: uuid
                nonce:
                    type: string
        PasskeyVerifyInput:
            type: object
            required: [sessionId, nonce, response]
            properties:
                sessionId:
                    type: string
                    format: uuid
                nonce:
                    type: string
                response:
                    type: object
                    description: Raw `AuthenticationResponseJSON` / `RegistrationResponseJSON` from `@simplewebauthn/browser`.
        PasskeyVerifyResult:
            type: object
            required: [verified, user]
            properties:
                verified:
                    type: boolean
                user:
                    $ref: "#/components/schemas/User"
        AuthenticationOptionsInput:
            type: object
            required: [username, sessionId, action]
            properties:
                username:
                    type: string
                sessionId:
                    type: string
                    format: uuid
                action:
                    type: string
                    description: |
                        Logical name of the action being approved, e.g. `transfer`,
                        `bill.pay`, `card.issue`, `password.change`.
                params:
                    description: Action-specific parameters (hashed server-side as `paramsHash`).
        AuthenticationOptionsOutput:
            type: object
            required: [options, paramsHash]
            properties:
                options:
                    type: object
                    description: WebAuthn `PublicKeyCredentialRequestOptionsJSON`.
                paramsHash:
                    type: string
        AuthenticationVerifyInput:
            type: object
            required: [username, response]
            properties:
                username:
                    type: string
                response:
                    type: object
                    description: Raw `AuthenticationResponseJSON`.
        AuthenticationVerifyResult:
            type: object
            required: [verified, actionToken, expiresAt]
            properties:
                verified:
                    type: boolean
                actionToken:
                    type: string
                expiresAt:
                    type: integer
                    format: int64
...
paths:
...
    /webauthn/registration/options:
        post:
            operationId: postWebauthnRegistrationOptions
            tags: [02-Identity-Auth]
            summary: First-time passkey enrollment — fetch creation options.
            description: |
                Fetches parameter structures for `navigator.credentials.create`.
                **Prerequisites:** `postIdentityLoginPassword` context.
            x-call-order: 3
            x-depends-on: [postIdentityLoginPassword]
            parameters:
                - $ref: "#/components/parameters/SessionIdHeader"
            requestBody:
                x-decrypted-request:
                    $ref: "#/components/schemas/PasskeyOptionsInput"
                description: |
                    Encrypted envelope (`EnvelopedRequest`). Plaintext is a `DecryptedEnvelope` whose `data` field conforms to `PasskeyOptionsInput`.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EnvelopedRequest"
            responses:
                "200":
                    description: WebAuthn creation options.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EnvelopedResponse"
    /webauthn/registration/verify:
        post:
            operationId: postWebauthnRegistrationVerify
            tags: [02-Identity-Auth]
            summary: First-time passkey enrollment — verify the attestation.
            description: |
                Persists credentials, sets state parameters flag.
                **Prerequisites:** `postWebauthnRegistrationOptions`.
            x-call-order: 4
            x-depends-on: [postWebauthnRegistrationOptions]
            parameters:
                - $ref: "#/components/parameters/SessionIdHeader"
            requestBody:
                x-decrypted-request:
                    $ref: "#/components/schemas/PasskeyVerifyInput"
                description: |
                    Encrypted envelope (`EnvelopedRequest`). Plaintext is a `DecryptedEnvelope` whose `data` field conforms to `PasskeyVerifyInput`.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EnvelopedRequest"
            responses:
                "200":
                    x-decrypted-response:
                        $ref: "#/components/schemas/PasskeyVerifyResult"
                    description: |
                        Credential registered.

                        Decrypted body is a `DecryptedEnvelope` whose `data` field conforms to `PasskeyVerifyResult`.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EnvelopedResponse"
...
    /webauthn/authentication/options:
        post:
            operationId: postWebauthnAuthenticationOptions
            tags: [03-StepUp]
            summary: Begin a step-up — fetch a challenge bound to (action, params).
            description: |
                Issues unique parameter challenges checking verification details.
                **Prerequisites:** Valid logged-in contextual session.
            x-call-order: 1
            x-depends-on:
                - postIdentityLoginPasskeyVerify
                - postWebauthnRegistrationVerify
            requestBody:
                x-decrypted-request:
                    $ref: "#/components/schemas/AuthenticationOptionsInput"
                description: |
                    Encrypted envelope (`EnvelopedRequest`). Plaintext is a `DecryptedEnvelope` whose `data` field conforms to `AuthenticationOptionsInput`.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EnvelopedRequest"
            responses:
                "200":
                    x-decrypted-response:
                        $ref: "#/components/schemas/AuthenticationOptionsOutput"
                    description: |
                        Challenge mapped. Decrypted body matches `AuthenticationOptionsOutput`.

                        Decrypted body is a `DecryptedEnvelope` whose `data` field conforms to `AuthenticationOptionsOutput`.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EnvelopedResponse"
    /webauthn/authentication/verify:
        post:
            operationId: postWebauthnAuthenticationVerify
            tags: [03-StepUp]
            summary: Complete a step-up — mint a one-shot action token.
            description: |
                Issues short-lived validation assertions tracking structural arrays.
                **Prerequisites:** `postWebauthnAuthenticationOptions`.
            x-call-order: 2
            x-depends-on: [postWebauthnAuthenticationOptions]
            requestBody:
                x-decrypted-request:
                    $ref: "#/components/schemas/AuthenticationVerifyInput"
                description: |
                    Encrypted envelope (`EnvelopedRequest`). Plaintext is a `DecryptedEnvelope` whose `data` field conforms to `AuthenticationVerifyInput`.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EnvelopedRequest"
            responses:
                "200":
                    x-decrypted-response:
                        $ref: "#/components/schemas/AuthenticationVerifyResult"
                    description: |
                        Verification calculated. Decrypted pattern conforms to `AuthenticationVerifyResult`.

                        Decrypted body is a `DecryptedEnvelope` whose `data` field conforms to `AuthenticationVerifyResult`.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EnvelopedResponse"

    # =========================================================================
    # 04-KYC-Customer
    # =========================================================================
...